Move Levenshtein extension to separate file
